+++ src/backend/app/Controller/LogsController.php
@@ -4,8 +4,16 @@ App::uses('AppController', 'Controller');
  * Logs Controller
  *
  * @property Log $Log
+ * @property PaginatorComponent $Paginator
  */
 class LogsController extends AppController {
+
+/**
+ * Components
+ *
+ * @var array
+ */
+	public $components = array('Paginator');
 
 /**
  * index method
@@ -26,7 +34,7 @@ class LogsController extends AppController {
  */
 	public function view($id = null) {
 		if (!$this->Log->exists($id)) {
-			throw new NotFoundException(__('Invalid log'));
+			throw new NotFoundException(__('Log invalide.'));
 		}
 		$options = array('conditions' => array('Log.' . $this->Log->primaryKey => $id));
 		$this->set('log', $this->Log->find('first', $options));
@@ -41,15 +49,18 @@ class LogsController extends AppController {
 		if ($this->request->is('post')) {
 			$this->Log->create();
 			if ($this->Log->save($this->request->data)) {
-				$this->Session->setFlash(__('The log has been saved'));
-				$this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('Log enregistré.'), 'flash/success');
+				$this->redirect(array('action' => 'view', $this->Log->id));
 			} else {
-				$this->Session->setFlash(__('The log could not be saved. Please, try again.'));
+				$this->Session->setFlash(__('Log impossible à enregistrer. Réessayez ultérieurement.'), 'flash/error');
 			}
 		}
 		$cards = $this->Log->Card->find('list');
 		$doors = $this->Log->Door->find('list');
 		$this->set(compact('cards', 'doors'));
+		foreach($this->passedArgs as $k => $v) {
+		    $this->set('default_'.$k, $v);
+		}
 	}
 
 /**
@@ -60,20 +71,22 @@ class LogsController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
+        $this->Log->id = $id;
 		if (!$this->Log->exists($id)) {
-			throw new NotFoundException(__('Invalid log'));
+			throw new NotFoundException(__('Log invalide.'));
 		}
 		if ($this->request->is('post') || $this->request->is('put')) {
 			if ($this->Log->save($this->request->data)) {
-				$this->Session->setFlash(__('The log has been saved'));
-				$this->redirect(array('action' => 'index'));
+				$this->Session->setFlash(__('Log sauvegardé.'), 'flash/success');
+				$this->redirect(array('action' => 'view', $this->Log->id));
 			} else {
-				$this->Session->setFlash(__('The log could not be saved. Please, try again.'));
+				$this->Session->setFlash(__('Log impossible à enregistrer. Réessayez ultérieurement.'), 'flash/error');
 			}
-		} else {
-			$options = array('conditions' => array('Log.' . $this->Log->primaryKey => $id));
-			$this->request->data = $this->Log->find('first', $options);
 		}
+		$options = array('conditions' => array('Log.' . $this->Log->primaryKey => $id));
+		$this->request->data = $this->Log->find('first', $options);
+        $this->set('log', $this->request->data);
+
 		$cards = $this->Log->Card->find('list');
 		$doors = $this->Log->Door->find('list');
 		$this->set(compact('cards', 'doors'));
@@ -83,20 +96,23 @@ class LogsController extends AppController {
  * delete method
  *
  * @throws NotFoundException
+ * @throws MethodNotAllowedException
  * @param string $id
  * @return void
  */
 	public function delete($id = null) {
+		if (!$this->request->is('post')) {
+			throw new MethodNotAllowedException();
+		}
 		$this->Log->id = $id;
 		if (!$this->Log->exists()) {
-			throw new NotFoundException(__('Invalid log'));
+			throw new NotFoundException(__('Log invalide.'));
 		}
-		$this->request->onlyAllow('post', 'delete');
 		if ($this->Log->delete()) {
-			$this->Session->setFlash(__('Log deleted'));
+			$this->Session->setFlash(__('Log supprimé.'), 'flash/success');
 			$this->redirect(array('action' => 'index'));
 		}
-		$this->Session->setFlash(__('Log was not deleted'));
+		$this->Session->setFlash(__('Log impossible à supprimer.'), 'flash/error');
 		$this->redirect(array('action' => 'index'));
 	}
 }
